Allow installer to continue without uninstalling previous version
Sometimes users have gotten into the situation where they could not install
Tiled because the installer detected a previous version, which could not be
uninstalled.

The installation of new versions now do not uninstall the previous version,
but overwrite thoes files.

The automappingconverter is being linked to from start menu as well.